refactor(routing): apply code review fixes for shared-state safety and SRP
Clone allow-lists when merging candidates so the cached read model is never
mutated across requests, report fallback provenance correctly on the LLM span
when the first pick comes from the fallback chain, skip candidate building on
the inline no-intent hot path, and extract the routing layer from the
forwarder into routing.go behind a routedBackend value object.
